Sarkars/merge utils (#33)
