+++ src/ngraph_utils.cc
@@ -14,7 +14,6 @@
  * limitations under the License.
  *******************************************************************************/
 #include "ngraph_utils.h"
-#include "ngraph_api.h"
 
 #include <fstream>
 #include <iomanip>
@@ -25,6 +24,7 @@
 #include "ngraph/distributed.hpp"
 #endif
 
+#include "tensorflow/core/common_runtime/dma_helper.h"
 #include "tensorflow/core/common_runtime/optimization_registry.h"
 #include "tensorflow/core/framework/attr_value_util.h"
 #include "tensorflow/core/framework/graph.pb.h"
@@ -34,6 +34,8 @@
 #include "tensorflow/core/platform/default/logging.h"
 #include "tensorflow/core/platform/protobuf.h"
 
+#include "version.h"
+
 using namespace std;
 namespace ng = ngraph;
 
@@ -41,8 +43,60 @@ namespace tensorflow {
 
 namespace ngraph_bridge {
 
+Status IsNgraphTFLogTensorCopiesEnabled(int graph_id,
+                                        bool& is_copy_log_enabled) {
+  const char* copy_env_var = std::getenv("NGRAPH_TF_LOG_TENSOR_COPIES");
+  if (copy_env_var == nullptr) {
+    is_copy_log_enabled = false;
+    return Status::OK();
+  }
+  int test_graph_id;
+  try {
+    test_graph_id = stoi(string(copy_env_var));
+  } catch (const std::invalid_argument& ia) {
+    return errors::InvalidArgument(
+        "Invalid argument for NGRAPH_TF_LOG_TENSOR_COPIES");
+  }
+  // if -1 copies are logged for all graphs
+  is_copy_log_enabled = (test_graph_id == -1 || test_graph_id == graph_id);
+  return Status::OK();
+}
+void PrintTFTensor(Tensor& T1) {
+  NGRAPH_VLOG(4) << "all tensor values" << (T1).SummarizeValue(64) << endl;
+}
+std::string DebugNode(Node* node) {
+  std::string temp = node->name();
+  temp += "[" + node->type_string() + "]";
+  return temp;
+}
+
+std::string PrintBool(bool var) { return (var ? "Yes" : "No"); }
+
 bool IsNGVariableType(string node_type) {
-  return node_type == "NGraphVariable";
+  if (ngraph_tf_are_variables_enabled())
+    return (node_type == "NGraphVariable" || node_type == "NGraphAssign");
+  else
+    return node_type == "NGraphVariable";
+}
+
+bool IsNGSupportedType(string node_type) {
+  return (IsNGVariableType(node_type) || node_type == "NGraphEncapsulate");
+};
+
+// Read from this ng_tensor into tf_tensor
+void ReadNGTensor(shared_ptr<ng::runtime::Tensor> ng_tensor,
+                  Tensor* tf_tensor) {
+  void* tf_src_ptr = (void*)DMAHelper::base(tf_tensor);
+  ng_tensor->read(tf_src_ptr, 0, ng_tensor->get_element_count() *
+                                     ng_tensor->get_element_type().size());
+}
+
+// Write into this ng_tensor from tf_tensor
+void WriteNGTensor(shared_ptr<ng::runtime::Tensor> ng_tensor,
+                   Tensor* tf_tensor) {
+  void* tf_src_ptr = (void*)DMAHelper::base(tf_tensor);
+  ng_tensor->write(tf_src_ptr, 0, ng_tensor->get_element_count() *
+                                      ng_tensor->get_element_type().size());
 }
 
 void SummarizeOp(OpKernelConstruction* ctx, std::ostream& out) {
